requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-cloud-bigquery
python-dotenv
gspread
google
loguru
pandas
numpy
pydantic
google-cloud-storage
msoffcrypto-tool
gspread-dataframe